requirements.txt
mysql-connector-python==8.0.33
protobuf==3.20.3
PyMySQL==1.1.0
pyodbc==4.0.39
